Update with ticket pass level by level with login viwe with login function including session validation

+++ PeojectTWI/Controllers/InventoryController.cs
@@ -28,21 +28,33 @@ namespace PeojectTWI.Controllers
             _warrentyService = new WarrentyService();
         }
 
+
         public ActionResult insertProductCategory()
         {
-       
+            if (Session["LoggedUserID"] == null)
+            {
+                return RedirectToAction("Login", "Home");
+            }
             return View();
         }
 
         [HttpPost]
         public ActionResult insertProductCategory(productCategory pd)
         {
+            if (Session["LoggedUserID"] == null)
+            {
+                return RedirectToAction("Login", "Home");
+            }
             _inventoryService.insertProductCategory( pd.BrandName, pd.VendorName, pd.VendorContact, pd.VendorEmail, pd.VendorAddress, pd.ProductName);
             return View();
         }
 
         public ActionResult viewProductCategories()
         {
+            if (Session["LoggedUserID"] == null)
+            {
+                return RedirectToAction("Login", "Home");
+            }
             var pdc = _inventoryService.viewProductCategories();
             //var pdc = db.tblProductCategories;
             return View(pdc);
@@ -50,6 +62,10 @@ namespace PeojectTWI.Controllers
 
         public ActionResult editProductCategory()
         {
+            if (Session["LoggedUserID"] == null)
+            {
+                return RedirectToAction("Login", "Home");
+            }
             var id = Request["pid"];
             int pid = Convert.ToInt32(id);
             var a = _inventoryService.getProductCategoryDetails(pid);
@@ -60,7 +76,10 @@ namespace PeojectTWI.Controllers
         [HttpPost]
         public ActionResult updateProductCategory(productCategory pdc)
         {
-
+            if (Session["LoggedUserID"] == null)
+            {
+                return RedirectToAction("Login", "Home");
+            }
             try
             {
                 _inventoryService.updateProductCategory(pdc.ProductID, pdc.BrandName, pdc.VendorName, pdc.VendorContact, pdc.VendorEmail, pdc.VendorAddress, pdc.Active, pdc.ProductName);
@@ -76,6 +95,10 @@ namespace PeojectTWI.Controllers
 
         public ActionResult deleteProductCategory()
         {
+            if (Session["LoggedUserID"] == null)
+            {
+                return RedirectToAction("Login", "Home");
+            }
             var id = Request["pidDelete"];
             int pid = Convert.ToInt32(id);
             var a = _inventoryService.getProductCategoryDetails(pid);
@@ -85,6 +108,10 @@ namespace PeojectTWI.Controllers
 
         public ActionResult RemoveProductCategory()
         {
+            if (Session["LoggedUserID"] == null)
+            {
+                return RedirectToAction("Login", "Home");
+            }
             var id = Request["pidDelete"];
             int pid = Convert.ToInt32(id);
              _inventoryService.deleteProductCategory(pid);
@@ -95,6 +122,10 @@ namespace PeojectTWI.Controllers
 
         public ActionResult insertMasterStore()
         {
+            if (Session["LoggedUserID"] == null)
+            {
+                return RedirectToAction("Login", "Home");
+            }
             masterStore mss = new masterStore();
             using (var context = new ProjectDBEntities())
             {
@@ -112,6 +143,10 @@ namespace PeojectTWI.Controllers
         [HttpPost]
         public ActionResult insertMasterStore(masterStore ms)
         {
+            if (Session["LoggedUserID"] == null)
+            {
+                return RedirectToAction("Login", "Home");
+            }
             _inventoryService.insertMasterStore(ms.ProductId, ms.perchesedCount, ms.unitPrice, ms.PerchesedDate);
             masterStore mss = new masterStore();
             using (var context = new ProjectDBEntities())
@@ -132,12 +167,20 @@ namespace PeojectTWI.Controllers
 
         public ActionResult viewtMasterStore()
         {
+            if (Session["LoggedUserID"] == null)
+            {
+                return RedirectToAction("Login", "Home");
+            }
             var pdc = _inventoryService.viewMasterStore();
             return View(pdc);
         }
 
         public ActionResult editMasterStore()
         {
+            if (Session["LoggedUserID"] == null)
+            {
+                return RedirectToAction("Login", "Home");
+            }
             var id = Request["msid"];
             int mid = Convert.ToInt32(id);
             var a = _inventoryService.getMasterStoreDetails(mid);
@@ -148,7 +191,10 @@ namespace PeojectTWI.Controllers
         [HttpPost]
         public ActionResult updateMasterStore(masterStore ms)
         {
-
+            if (Session["LoggedUserID"] == null)
+            {
+                return RedirectToAction("Login", "Home");
+            }
             try
             {
                 _inventoryService.updateMasterStore(ms.mStoreId, ms.ProductId, ms.perchesedCount, ms.unitPrice, ms.PerchesedDate, ms.RecoredEnterdBy);
@@ -164,6 +210,10 @@ namespace PeojectTWI.Controllers
 
         public ActionResult deleteMasterStore()
         {
+            if (Session["LoggedUserID"] == null)
+            {
+                return RedirectToAction("Login", "Home");
+            }
             var id = Request["msIDDelete"];
             int mid = Convert.ToInt32(id);
             var a = _inventoryService.getMasterStoreDetails(mid);
@@ -173,6 +223,10 @@ namespace PeojectTWI.Controllers
 
         public ActionResult removeMasterStore()
         {
+            if (Session["LoggedUserID"] == null)
+            {
+                return RedirectToAction("Login", "Home");
+            }
             var id = Request["msDeleteid"];
             int msid = Convert.ToInt32(id);
             _inventoryService.deleteMasterData(msid);
@@ -181,12 +235,17 @@ namespace PeojectTWI.Controllers
 
 
         public ActionResult WareHouseManage()
-        {     
+        {
+            if (Session["LoggedUserID"] == null)
+            {
+                return RedirectToAction("Login", "Home");
+            }
             return View();
         }
 
         public JsonResult GetProductCategoryValues()
         {
+
             var ProdcutCategory = _inventoryService.GetProductCategoryValues();
             return Json(ProdcutCategory, JsonRequestBehavior.AllowGet);
         }
@@ -198,6 +257,7 @@ namespace PeojectTWI.Controllers
 
         public JsonResult CheckExist(int dataType, int ProductCategory, string SerialNo)
         {
+
             var ExsistReslut = _inventoryService.CheckExsist(dataType, ProductCategory, SerialNo);
 
             return Json(ExsistReslut, JsonRequestBehavior.AllowGet);
@@ -206,6 +266,7 @@ namespace PeojectTWI.Controllers
 
         public JsonResult ManageWareHouseData(int dataType,int ProductCategory,  string SerialNo)
         {
+
             var whmanageReslut = _inventoryService.ManageWareHouseData(dataType, ProductCategory, SerialNo);
 
             return Json(whmanageReslut, JsonRequestBehavior.AllowGet);
@@ -215,11 +276,16 @@ namespace PeojectTWI.Controllers
 
         public ActionResult ViewWareHouse()
         {
+            if (Session["LoggedUserID"] == null)
+            {
+                return RedirectToAction("Login", "Home");
+            }
             return View();
         }
 
-        public JsonResult ViewWareHousedata(DateTime? FromDate = null, DateTime? Todate = null) 
+        public JsonResult ViewWareHousedata(DateTime? FromDate = null, DateTime? Todate = null)
         {
+
             List<WHmanage> WHList = _inventoryService.ViewWareHousedata(FromDate, Todate);
             return Json(WHList,JsonRequestBehavior.AllowGet);        
         }
